+++ Source/KSCrash/Recording/Tools/KSString.h
@@ -34,7 +34,6 @@ extern "C" {
 
 #include <stdbool.h>
 #include <stdint.h>
-#include <sys/types.h>
 
 
 /** Check if a memory location contains a null terminated UTF-8 string.
@@ -45,9 +44,7 @@ extern "C" {
  *
  * @param maxLength The maximum length to be considered a valid string.
  */
-bool ksstring_isNullTerminatedUTF8String(const void* memory,
-                                         int minLength,
-                                         int maxLength);
+bool ksstring_isNullTerminatedUTF8String(const void* memory, int minLength, int maxLength);
 
 /** Extract a hex value in the form "0x123456789abcdef" from a string.
  *
@@ -59,9 +56,7 @@ bool ksstring_isNullTerminatedUTF8String(const void* memory,
  *
  * @return true if the operation was successful.
  */
-bool ksstring_extractHexValue(const char* string,
-                              size_t stringLength,
-                              uint64_t* result);
+bool ksstring_extractHexValue(const char* string, int stringLength, uint64_t* result);
 
 /** Replace a string with another string.
  * If dest points to a non-null, it will be freed.
